Enable selection of init by api trace
Due to the feature implemented in
I29c815306418e7511b7ffef95dee54f6c1da894d,
where selecting an already selected artifact would
be do nothing, instead of reparsing. However, in
the case of the api initiated tracing, the artifact
was generated if the user was in monitor stage or
the capture stage, the internal state of selected artifact
would be updated with this new trace, but the ui would not
change stages to the new capture. Thus, the ui and the
state of the "selected" artifact were out of sync.

This change fixes this issue by only setting an api
initiated trace as the selected artifact if the cpu
profiler enters the capture stage.

Fixes: 261073930
Test: none
Change-Id: If5f233d066e5a8bc459c47393012e75e734f6704

GitOrigin-RevId: 9c6692804ca32e9afe559f0e2a64b02e30e84f30