Complete and fix BasicLayoutEditTest
This change updates BasicLayoutEditTest so it matches the specification
from Test Tracker UUID 78baed4b-32be-4d72-b558-ba9f6c727334. The test
was slightly modified to meet the spec: we're checking the text value of
a different TextView, and waitForAllBackgroundTasksToBeCompleted() had
to be called in order to guarantee the new text will be applied by the
time we verify it.

Bug: 254435642
Test: BasicLayoutEditTest
Change-Id: Iea13033b056bffb19cc9b4c1ebb4ee6629faf000

GitOrigin-RevId: 4a38a3d0a7a78c100a006ad04d08540cd6bf09b2