Change strategy for finding library resource classes
Instead of checking the last build status as a proxy for whether the
project system class loader will be able to load resouce classes,
inquire directly by using the project system class loader's
loadClass() method.

ProjectSystemClassLoader is a DelegatingClassLoader.Loader, so its
loadClass() method does not define a class; this has the benefit of
being side-effect-free, but on the other hand it means that in the
case of final IDs being used we will load the class bytes twice.

This is also still vulnerable to a (small) TOCTTOU window, where a
file modification or deletion occurs between the check in
LibraryResourceClassLoader and the eventual load in the
ModuleClassLoader.  Rethinking and rearranging the order of class
loaders and their contracts could probably remove this TOCTTOU window.

The delegation to built artifacts when final ids are in use is I think
counter to the tests in ModuleClassLoaderTest related to priority
between R class files and AARs, which was previously not exposed
because those tests did not perform a build.  Force those tests to use
non-final ids.

Bug: 247072439
Test: existing tests adjusted
Change-Id: Iba9174b567fc0ed6464e595b1acb1a74f44f9586

GitOrigin-RevId: aaa7759bc152ac9f97b7e2d8d8e2d67285ea6106