Run activity safe delete before cmake
The CMake safe delete processor was running before ours and "stealing"
the activity file from us before we had a chance to process it. By
positioning our processor first, we ensure that doesn't happen. See also
https://youtrack.jetbrains.com/issue/CPP-31462 filed for CLion to fix
that behavior.

Unfortunately I can't find a good way to test this.
`LayoutBindingSafeDeleteProcessorTest` actually tests this exact
scenario; but since it's in the data binding module, which doesn't
depend on anything related to C++, it doesn't run with the CMake
processor installed. I'm not aware of any way to have injected this
error short of a full integration test, and I don't think the issue
warrants that.

Fixes: 261154541
Test: Manually verified
Change-Id: Idf4eaa393e2edfac6d57eab631abf934193189c4

GitOrigin-RevId: 38e87e18d4dd08871e61ca2d6b07adae0f15ebbb