Device-side support for stopping/starting video streaming
The StartVideoStreamMessage and StopVideoStreamMessage control messages may be used to start/stop video streaming without stopping the agent.
The START_VIDEO_STREAM flag can be used to start the screen sharing agent with or without starting video streaming.

The Studio side of starting/stopping the video streaming will follow in a separate CL.

Test: tests will be added when the Studio side is implemented
Bug: 247884979
Change-Id: I00d3a021b4f33336db6cf3c061d7eeda8c0551db

GitOrigin-RevId: 7297e80526e3a19d4ad21ce25c32be07c3c157df