Only transition single device to DISABLED
When listening for ADB events, only transition LiveEdit state for the
specific device where the application disconnects.

Right now, if you are running on multiple devices, and the app closes on one of those devices, all device states will incorrectly transition to "DISABLED" and the LiveEdit icon will disappear. This CL resolves that by ensuring the transition to DISABLED only occurs on the device where the app actually closed.

Bug: None
Test: Manual
Change-Id: I34ec131dc230894ce9b3618ed28a10f1657255e9

GitOrigin-RevId: aaf8d9358b2f27811ef47c7806d0c772dba94896