[AUA] Workaround for missing entries in indexes
Under some circumstances, the PropertiesFile method
findPropertyByKey() can query indexes for the properties of
gradle/wrapper/gradle-wrapper.properties, with those indexes not
reflecting the contents of the files.  (This was seen recently after
using the Gradle plugin's quickfix to update the version of Gradle
after use of a JDK incompatible with the current version of Gradle).

Workaround this unexplained behavior by explicitly listing and
filtering the properties in the wrapper file ourselves.

Bug: 251266381
Test: manual
Change-Id: I9aa6d4e7adf44b9e80327eb7e84418cfc9fe2a94

GitOrigin-RevId: 11cc11f7800420b43b26004d6bc0e7bc5315644f