Fix sometimes zoom buttons don't work
The setScale() function denies the change of zooming when the
difference between current and new scales are smaller than tolerance.

Functions canZoomIn() and canZoomOut() should also consider the
tolerance, otherwise the zooming functions may do nothing if the
difference of current and new scales are smaller than the tolerance
value.

Bug: 261715145
Test: Added
Change-Id: Ifd64ba44e3130b233db62ca2637f3441b2bd8a8e

GitOrigin-RevId: 0c0c609b1edfbf1dee8d735be29764ee43d622c8