Support optional bootClasspath libraries
Android SDK libraries from `optional` directory are not set up as a part
of an IDE Android SDK entity and needs to be added to IDE modules as
regular dependencies.

Recognize Android SDK libraries from the `optional` directory and
add them to modules as regular dependencies. (Initially add support for
AGP versions supporting V2+ models.)

(1) Propagate `bootClasspath` to the place AGP models are converted to
    IDE library and dependency models;

(2) Recognize those `bootClasspath` entries that refer to libraries in
    the `optional` directory;

(3) Create `IdeJavaLibrary` for these entries and add them as
    dependencies to all artifacts of the Gradle project.

Bug: 234749386
Test: testSimpleApplication_withAndroidCar
Change-Id: Ife70d84b8765a0fc205e6a0dc7dba683d2488ab7

GitOrigin-RevId: 81dda2989cc5d3907fe2a924ddf74dd7699dc046