Adapt to new Start/Stop Heap Sample commands
The start and stop heap sample command proto
messages have been converted to have the same
fields as the start and stop cpu trace commands.
As such, we send in different types of data into
the command now for heap sampling. The command
handlers have been updated respectivel in a base-side
change.

Bug: 263172101
Test: existing
Change-Id: Id92d31a0f170ffcc82258f2f52de717785d2e7ed

GitOrigin-RevId: aca3c83f2fe41d65b33d4eb2a8dd33889fc1d621