Fix DnD behavior
Before this change if the insertion point was one of the items
moved/copied we would ignore this insertion point and choose an
insertion point further down in the tree.

This logic was flawed. One example:
If the previous item could accept items there is no need to adjust
the insertion point.

After this change:
- remove the insertion point adjustment during drop
- add a check in update to see if the insertion point is valid
  for the found receiver. If not disallow the drop.
- if this is a move: do not allow a move to the same spot.

Bug: 210465107
Test: TreeTableDropTargetHandler
Change-Id: I666816fc5e1ebf4df397d141fba1e3bdb22e248d

GitOrigin-RevId: 348a85a528a6da7878d09cc11da8c01e93c7f329