Make app insights code sharable
This change follows up on the previous one and updates the files in place to make them usable. Some of these classes have been made generic in an attempt to make them reusable in the future when we integrate play vitals API.

They are:
1) AppInsightsState
2) AppInsightsModuleController

TimeIntervalFilter and Fatality types are moved to the new module. They are made to include all possible values used by both CL and Vitals.  Users can choose which values they want to use.

Test: n/a
Bug: 263259654
Change-Id: I48ba13ffc4d21af6cab46ee554d48e5dc13bd671

GitOrigin-RevId: 2cd1341a0d37379e3e72d982372e16a2f18e8b73