Add missing flags needed to start Game Tools
In order to start game tools, two new options needed to be passed with
game tools:

  -Djava.system.class.loader=com.intellij.util.lang.PathClassLoader
  -Didea.vendor.name=Google

In order for PathClassLoader to work, we need to set the vendor name. If
we don't, start-up will fail with an invalid state exception as IntelliJ
will start another activity in an attempt to figure out the vendor name.

Bug: 263887432
Change-Id: I02277d9d3450b93e064d7d6b3f345fd33ba98938

GitOrigin-RevId: 5ddb146f2c711417cdbacc58417ab0387de323fb