Ignore non-final lambda in Compose autocomplete
When a @Composable function's final required argument is a lambda with no arguments, and there are optional arguments that follow it, we are incorrectly giving autocomplete suggestions with a trailing lambda that is invalid.

As an example, with this function:
  @Composable
  fun Foo(reqArg: @Composable () -> Unit, optArg: Int = 0) {}

we currently propose the following completion:
  Foo() {
  }

That completion is invalid. Instead, we should propose:
  Foo(reqArg = { /*TODO*/ }


Fixes: 209060418
Test: ComposeCompletionContributorTest
Change-Id: I423cc2ee049c52bf69f311058f1d885d5dbe7919

GitOrigin-RevId: fc0dd8f8fbf21530a672b9cf4d9ab80620a3b87a