Fix filtering of project configurables
AndroidStudioPreferences was using
ExtensionPoint.unregisterExtensions(Predicate) to filter out unwanted
project configurables. Following the removal of that method,
Change I48e8946c updated AndroidStudioPreferences to use
unregisterExtensions(BiPredicate, boolean) instead. However, the
predicate of the latter method has the opposite meaning compared to the
predicate of the former method, which was not taken into account when
the change was made.
This led to unregistering the project configurables that were supposed
to be kept and keeping the ones that should have been removed.

Bug: 264434280
Test: N/A
Change-Id: Ibcda5c864cc18e7bf516261c99d8c3a9951e9afb

GitOrigin-RevId: 5081516978f530561b9a3fa720afbb1759fc8a0f