[Dsl] Handle indirect settings of rootProject attributes
In particular, setting rootProject.name in the root settings file ends
up inserting a property in the root build file, as we resolve
references (and lvalues) to rootProject there.  This property
insertion was previously automatically added as EXISTING,
corresponding to an entry textually present in the build file, which
was unfortunately incorrect, leading to errors when attempting to add
blocks with required positioning (such as requiring the `plugins` block
to be before any `ext` block).

Detect whether -- after lvalue resolution -- we are adding a property
to a file distinct from the file being parsed.  If so, add it to the
original file in HIDDEN state, so that it is recognized as being
physically present but does not contribute to that file's Dsl property
model; then add it to the resolved block in APPLIED state, so that it
does contribute to the property model but is not treated as being
textually present.

Fixes: 264016823
Test: manual, added
Change-Id: I39d441950aaa0e3c04c7f233fc9ce748097ae7cf

GitOrigin-RevId: f227f615ef55a243dffb34e465606fef1c95208c