Fix ModuleClassLoaderOverlaysTest after 223.4884.69 update
After the update, UrlClassLoader will cache the directory structure on
creation. The test was creating the loader before the .class files were
being added and the UrlClassLoader would cache the empty tree structure
causing the classes not to be found later.

In production code, the overlay is not pushed until after the classes
have been added so the test was testing a case that would not happen
causing the test to fail.

Fixes: 263894245
Test: ModuleClassLoaderOverlaysTest
Change-Id: Ia820e3860b03e9fd83363855063621117db71748

GitOrigin-RevId: 2967c4705f9e2f0108b5691ed299c09a8f1c17e2