Deduplicate Compose compiler in IDE plugin
The Compose IDE plug-in ships with a build of the Compose
compiler plugin in order to register the compiler extensions
for the compiler used in the editor.

The compiler extensions can be directly "re-exported" by the
plugin description.

This change furthermore sets us up to eliminate uses of the old
frontend in the compose-ide-plugin module. Remaining usages
appear to mostly be a use of the hasComposableAnnotation check,
that should be fairly straightforward to port to the Analysis
API.

Bug: 254509778
Bug: 265241076
Bug: 265250102
Test: //tools/adt/idea/compose-ide-plugin:intellij.android.compose-ide-plugin.tests_tests
Change-Id: I2564eb79bb42c9b5eb258b0fa9aa3c14ac01e140

GitOrigin-RevId: 95d09687388cde8f38c29763de460e2962217969