Fix race condition when an application is rerun
If the application is Rerun, we set the stated to Loading.
However, Rerun also terminates the app, which results in an app
termination signal from ADB. This signal ripples through the
DeviceEventWatcher and ends up setting the state to Disabled,
which prevents the state moving to UpToDate when the app loads.

The fix is to just check if the state is Loading, and if so, do
not set the state to Disabled. This is correct because when in
the Loading state, we have never recognized a Client, implying
we could never have a Client termination to begin with.

Bug: N/A
Test: manual
Change-Id: I455644cb55cbe29ad9db84fface16406478b8dd0

GitOrigin-RevId: 28584fcaa0f3f7a20202d4fe15c1c80ead57d145