Add missing power rails
Previously, the constraint to filter for power rails using
the power rail prefix of "power.rails". However, it was found
that 3 of the power rail names only have the prefix of "power.".
Consequently, these 3 power rails were excluded from the displayed
data. This change accounts for this exceptional case and aligns the
missing 3 power rail names with that of the perfetto UI as done with
every other rail previously when added.

Bug: 254116554
Test: manual
Change-Id: I0da5c842f32c1259ecf5c75b67ddc89aeb51a8f7

GitOrigin-RevId: 1c7457b1ef24e2d34a18e905b0569483d944e2f9