Make AndroidModuleInfo an interface
Rendering pipeline currently depends on AndroidModuleInfo. We are
trying to make rendering pipeline independent of studio and it
seems impossible to reuse current AndroidModuleInfo implementation
for it heavily depends on studio-specific object, such as
AndroidFacet, AndroidModel etc. Thus we make AndroidModuleInfo an
interface renaming the previous implementation to
StudioAndroidModuleInfo and allowing different implementations
to be plugged in to rendering pipeline through AndroidModuleInfo
interface.

Bug: 265765901
Test: all tests should pass
Change-Id: Ied34fd0782602298fe2e29d33e274adcfc642678

GitOrigin-RevId: 7e17d41569e30ab5001cc503d22836198e7dfb43