Implement DaggerDataIndexer
The indexer actually is what traverses the PSI tree and builds up the
list of index entries for a given file. This change adds an indexer for
Dagger utilizing the concepts and wrappers already added, so that this
file is focused largely on the actual traversal and what elements it's
interested in looking at.

I'm moving DaggerConceptIndexer[s] into DaggerDataIndexer.kt as well,
since it seems like a better home for the interfaces.

Bug: 265846405
Test: Included
Change-Id: I05729e97ff2e6493c7f9e6386c388d43c494115f

GitOrigin-RevId: 0cdaa76bb4f3c121e00f1a967a4f0ba4bc1b158d