Add DaggerIndex
All the pieces that make up this index already exist, so this is largely
just wiring them up.

Tests are only meant to validate that the index is working correctly,
and not to exhaustively verify all the different types of index entries.
Those are already heavily unit tested, so additional tests are not
appropriate here.

Bug: 265846405
Test: Included
Change-Id: I5453fd4c040d1c13b4d47a80712f9b18e6f5cf8e

GitOrigin-RevId: ed6a7c5a7c7e6d92ac12650f0195a4568d2d1011