Update the progress in stats
The analytics data for session attach has an error_info with an
attach_error_state, that was intended to hold an indication of
how far in the attach attempt we got stuck.

However the code only updates this field when we hit an explicit
error. Change this such that we know how far an attach session
went before it failed.

After this change the analytics query will still say:
- session.attach.error_info.attach_error_state
when in reality this is the last progress detected.
At this point I don't see a great reason for changing the proto.

Bug: N//A
Test: N/A
Change-Id: Ia3203ee854aacb1f229f89b04612d5ebd953dbdb

GitOrigin-RevId: de518d02f6a146a568955115ff7355b0fcfcb121