[Dsl] Improve model isModified()
Unlike the element layer, in the model we attempt to perform some kind
of semantic understanding of modification, trying to return false if
the element layer has been modified but to an equivalent value.

The implementation, however, did not handle modification in model
properties with transforms terribly accurately.  Compound elements
corresponding to multiple model properties (e.g. the infix form of
plugin application) would return true for isModified for all of those
properties if any one was modified.  Enhance the entry point to
consider elementwise modifications to accept original and new
instances of both raw elements and of transformed ones, and compare
the names of the former and the values of the latter.

Bug: 263584928
Test: existing test strengthened
Change-Id: I6f8f323af9481f9ed2ae93480168fe3f2fd84620

GitOrigin-RevId: 5c76f44d1b5047da4ba822968ea53ab97e3165ca