Hide tabs split controls for Java/Kotlin files if not needed
We open Java/Kotlin files with SeamlessTextEditorWithPreview in Android
Studio. IntelliJ new UI includes new split controls that are placed next
to the editor tabs, but they should not be displayed for Java or Kotlin
files that are not part of a Compose project and/or don't have custom
views in them.

This change creates a wrapper around getTabsAction() to make sure that
we hide the new tab split controls when they are not needed.

Bug: 266615219
Test: Manual
Change-Id: I504dfe5047875189ad1de62dfdad1d5b24fadfb6

GitOrigin-RevId: b5b6c7adc5546e6dd8b62cfbe688cc2952b3034d