Use the real DPI in Compose Preview with reference devices
When sizing the Compose Preview based on DP size instead of PX size, use
the real device DPI instead of the one coming from the definition. The
density in the definition of reference devices was chosen so that sizes
would look reasonable when displayed in PX, so that is not necessary
when using DP size.

Bug: 260234497
Test: N/A
Change-Id: Ic5d1c2d70860076fc5767cedc2876050ad14f7ec

GitOrigin-RevId: a6c1e52e174a70cb3149b21aefda8cbfbbac4317