Fix InformationPopup to not cancel erroneously
Problem: InformationPopup cancels (closing) when the mouse is over
a sub popup (popup within the InformationPopup) that is not within
the whole bounds of the InformationPopup.

Solution: Instead of checking against InformationPopup's own bounds,
we check against all windows owned by the popup (including child
popups). We also perform a test against the owner button that owns
the InformationPopup, as well as specially handling the case where
the user has to move the mouse out of the bounds of the owner button.

Supplement: Also accounts for the gap between the owner button and
the InformationPopup.

Fixes: 266720937
Test: InformationPopupImplTest
Change-Id: I4c198ca571107358ba79246f6006a623f9ceedb4

GitOrigin-RevId: b1a9f2eb859a4f476cbde586251ea77a888cfb0a