Fix new split controls not showing up sometimes
When setting the visibility of the new controls to true, we were also
checking if there were actions added to the action group. Since actions
are not updated immediately, There was a timing causing the children to
still be present in the group by the time the check was made, making us
not able to display the split controls in this case.

Fixes: 267294095
Test: Manual
Change-Id: Ie0268b42e7bbbd80b3c5de46b2593be49cb8ff63

GitOrigin-RevId: 0d3f798f0e78317f86a5a123234f35f3d0c0fa6e