Fix Finding of Receiver Variable
ComposeValueContributor.contributeValues() is called by KotlinStackFrame
with:
```
variables == thisVariables + otherVariables
```

So, it's not necessary to sort `variables` in order to find the receiver
variable. In fact, sorting doesn't actually work in some cases.

Also, name the variable after it's type rather than it's name because
the name is sometimes cryptic (for example, "FooBar_u24lambda_u2d0")

Bug: 259610488
Test: Manually
Change-Id: I87caa536b860e617ea55d0dac574a2a8b0877163

GitOrigin-RevId: d41a8985472cfe72de111bcc6c3fc2430a9261cf