[Dsl,.toml] Change the representation of type-safe aliases
The intermediate holder values generated as part of alias groups are
unlikely to be useful to users, and break the expected mapping of
property name to property.  Rather than computing effective maps to
hold those alias groups in `mapAliasesToAccessors` at parse-time,
preserve the Toml map structure directly.

This means that we must do more work at element resolution time: we
must search through all the properties of the table to find one that
matches the possibly-compound reference parts.  We must also perform
alias to accessor conversion ourselves at write-time, if we are
writing a reference to a property in a Version Catalog file into a Dsl
script file (Groovy or Kotlin).

This all means that we can now express more naturally the construction
of references to entities defined in Version Catalog files: we need
only look up the entity in the appropriate table by its key, not its
type-safe alias.

Bug: 200280395
Test: existing test adjusted
Change-Id: Id3ec9701d8c2b849e7dd0339bc5cdf076c2ccbf1

GitOrigin-RevId: 16316db638f508eac5709682ff78ba17b9013532