Add CaretPolicy to JTextPane used for Assistants
If the policy is not explicitly set to not update, it will queue events on the EDT to update the view to where the text was inserted. This causes the scroll panel to initialize towards the bottom when a TutorialCard is loaded lazily (based off where the latest text was inserted), or when opening directly to a TutorialCard.

Fixes: 262001187
Test: Manual Validation
Change-Id: I7add4e2ca969eb1b0e06af6651543e9c63e9e6d4

GitOrigin-RevId: b7fbe0311714deb3462230e935481d34a171223f