Replace com.intellij.util.containers.toArray with toTypedArray
The com.intellij.util.containers.toArray extension function is identical
to toTypedArray with one exception - if the argument is empty and the
size of the collection being turned into an array is empty, it will
return the argument. This is a questionable optimization in all but
the tightest of loops, but in cases where we allocate a new array
to make the call, we gain nothing but lose readability.

Bug: N/A
Test: None
Change-Id: I189eb2a0e1ad59a697afd69c093f04a5562bc20b

GitOrigin-RevId: c2b4d3e9e86b5fbd0b506591ed1120f8638470ef