Fix Render error when using an AbstractComposeView
Layoutlib can't render custom views that extends
AbstractComposeView because of ViewLifecycleOwner gets null value.
ViewTreeLifecycleOwner is not instantiated on time. When trying
to show the Preview the Lifecycle gets a null value throwing a render
error with the message: "ViewTreeLifecycleOwner not found".
The solution is to instantiate a fake LifecycleOwner when needed by
CustomViews implementing the AbstractComposeView.

Test: FakeSavedStateRegistryLoaderTest, ViewTreeLifecycleTransformTest
Bug: 187339385
Change-Id: I77f7994b829ea59d0b49b9adc27a4e484e9cccde

GitOrigin-RevId: 37e22e1f7f3f396448e5e6fe77df6ea938bdc6ff