[AUA] First draft of notes
Ideally I'd like these notes to live alongside the source code,
double-ideally with links to the relevant bits of the source code, to
act as overview documentation.  Failing that, this could become some
kind of Doc, which would make it easier to refer to other internal
artifacts / design documents, but much, much harder to keep even
vaguely in sync with the code.

Bug: n/a
Test: n/a
Change-Id: Id60d2ea2bb7f0ec000badd59026b002a82bfdf86

GitOrigin-RevId: 97ca212bcc0b167f14059c82bf6af6760a94f1d8