[Dsl] Allow writing infix elements with an initial method call
Writing infix elements using our representation involves writing the
first element of the infix element first, which does not actually
yield an infix element in the Psi, and so needs a certain amount of
special-casing.  The special-casing was already implemented when the
first element was a GradleDslLiteral (e.g. `id 'foo' version '1.2.3'`)
but for plugins we will need to be able to write infix elements with a
method call first (e.g. `alias(libs.plugins.foo) apply false`).  Add
some more special-case to both Groovy and Kotlinscript writers.

Bug: 267607937
Test: in later CL
Change-Id: Ic6766980854762685bcfeabcc66c0105abf4bfde

GitOrigin-RevId: 3f32e16ef39e629c18f8c066aa6172fd843d8726