Resolve related dagger elements using index
Resolving dagger elements requires a bit of munging of type names, since
the index is built without full resolution and might store a given type
in a variety of ways. This change adds the logic to figure out all the
various keys we need to look for when resolving a particular type.

I've also added the utility method that does the index lookup and
filters appropriately for a given element, using other logic that's
already in place. That `getRelatedDaggerItems` is unfortunately not
tested yet, because it requires the index to be installed and enabled
(which it's currently not). But the method is relatively simple calling
into existing functionality (which is all extensively unit tested), and
I've manually verified that the method works as expected when the index
is enabled. This will be tested by more e2e scenarios once the index is
enabled.

Test: DaggerIndexTest
Bug: 265846405
Change-Id: I1d4a3005bb18f7074fa4fa7ff847d22e51663a52

GitOrigin-RevId: 73e7ab3c2948de9d99b8b1abdc2da7c7aa0bfc22