Migrate project away from #USE_PROJECT_JDK and undefined gradleJvm
When gradleJvm isn't present in the .idea/gradle.xml by default
the projects are using the #USE_PROJECT_JDK, macro as the name
indicates requires of the project jdk to trigger sync located
under .idea/misc.xml. This fact is causing to break the
unidirectional flow regarding the jdk selection, more in particular
in case of failure the gradleJvm is updated with embedded jdk that
end up updating also the project jdk.

In order to keep a unique source of truth, we are going to migrate
those projects to a vendor-version i.e. jbr-17 standard when
possible.

Bug: 214428183
Test: MigrateProjectGradleFromMacrosJdkIntegrationTest
Change-Id: I8987becef91c4a43436d2217e0f0c74901de37b6

GitOrigin-RevId: bf4de91eb85faaa960a319a3ed500cfaad83788c