Support new #GRADLE_LOCAL_JAVA_HOME macro and use it for new projects
A support has been added in order to be able to validate in the jdk
integration tests this new way of specifying the gradle jdk to
trigger sync, this also includes assertions of the same.

Now this gradleJvm macro is the default for all new created projects
meaning that this simplifies our tests that doesn't require any more
to have a valid jdk.table entry in order to trigger sync, getting
the path from the java.home field under .gradle/config.properties.

Bug: 214428183
Test: ImportProjectWithoutIdeaJdkIntegrationTest
Test: GradleSyncUseLocalJavaHomeIntegrationTest
Test: GradleConfigPropertiesTest and existing
Change-Id: I6024d30eb2496fd43ad90afb182e5565cdf68740

GitOrigin-RevId: 40f200b7b3cef37e634fb7396fdefc952024526d