[Dsl] Make SdkOrPreviewTransform less magical
If (for example) compileSdkVersion().setValue(...) is given a
ReferenceTo, one possible intent of the user is that the value at that
reference will be changeable.  If so, we should probably not use the
type-specific operators -Sdk or -SdkPreview, as that would prevent
changing the value between released (integer) sdk and preview (String,
"android-") forms.  Use the generic setter in those cases, reserving
use of the -SdkPreview operator to when the value is a literal String
beginning "android-".

(This also results in some changes to tests where the Sdk value is
being set to a literal String not beginning "android-".  Probably
using these values would lead Sync or Build to fail, but at least the
Dsl model remains somewhat consistent.)

Bug: 267259672
Test: existing tests adjusted
Change-Id: I3d650b75163bab6ff71b146a01b5cb48ae52700b

GitOrigin-RevId: f72e80aaa4fe39ea78419b87bbf9eb336712b5c2