[BA] prepare to configuration cache becoming stable
Gradle is promising to make configuration cache stable feature
in version 8.1. In this change we are checking what Gradle version
is used in the project and remove ui references to CC as incubating
feature if version is equal to 8.1 or higher.
Note that we do not have version of gradle from build data so we are
relying on what Android Studio believes should be based on the settings
which can be inaccurate in some corner cases.

Bug: 243175483
Test: parametrize analyzer unit test to cover both logic paths,
add check to UI creation test

Change-Id: I180e8ecc60cdcd7bef9df9919fc5b4dfa610353f

GitOrigin-RevId: d053af9a73426508fa7b090673080da62e5d26bd