Allow for runtime dependencies to be null
This change allows the extension of the ModelBuilderParameter to enable
the IDE to chose whether or not the variant dependencies model should
contain the runtime classpath or not on a per project basis.

This change currently maintains the old behaviour in all cases, it just
ensures the parameter is set and that the return type for the runtime
classpath is not nullable.

Bug: N/A
Test: existing snapshots should cover
Change-Id: I46a3ab563ef3e1afe9cb8cfab726de86eb6f23b8

GitOrigin-RevId: b1d7051a21093ea4ace937896bb8d19e7f0071c2