Allow setting max cache size for ProjectSystemClassLoader
We were hard coding the max cache size of ProjectSystemClassLoader's
jar loader to ~1MB. This means that loading classes from bigger jar
files, e.g. render jar in ASwB, would result in a cache miss, and the
cache loader being executed for every loadClass call. When prefetching
all files, that means we were reading the entire jar everytime we wanted
to load a single class, defeating the purpose of the cache.

This change adds a new flag (PROJECT_SYSTEM_CLASS_LOADER_CACHE_LIMIT) to
allow controlling the max cache size, and set its value to 1GB in the
ASwB run configuration. The flag default value is ~1MB to keep the
previous behavior in non-ASwB builds.

Bug: 265935125
Test: ProjectSystemClassLoaderTest
Change-Id: Ic5b05fe5cd374dee2c1fe5c54517882e662b3afb

GitOrigin-RevId: 5178e11d23b9bc8ff946dd733d0850a7103f3125