Allow to ignore the 'selected' DrawState in SceneComponent
In SceneComponent, the SELECTED DrawState has priority over all
the other states, but this may not be always the expected behaviour.
For example, we want to add some selection logic to Compose Preview,
but we don't want the SELECTED state to have priority over the HOVER
state.

This change adds a flag into SceneComponent that will later allow
for SceneComponent providers to turn off the priority that the
SELECTED state has over the rest.

Bug: 232719077
Test: added
Change-Id: Ic0f8091c77eb9196af3026119d8c19df7186b605

GitOrigin-RevId: a2716676b767e84a3e6e6f58917f7f8ee201aff5