[BA] Add support for CC property without 'unsafe'
Configuration cache is becoming stable in 8.1 and gradle team says
'.unsafe' part will be removed from the property name.
Old property will stay for a while as an alias and will emit warning.
For us this means we need to be able to read both properties
and write the new property when we believe this feature is stable.

Bug: 243175483
Test: updated ConfigurationCachePropertyAccessTest.kt
Change-Id: I538ccf61ff53c2c02b2c478c84f2096f4837aeae

GitOrigin-RevId: 809595b7727b1bf811ba956268acbf2bf9be3d9a