Implement Dagger-related extensions with index
V2 of the Dagger-related extensions use the Dagger index. Since the
index and its associated utilities already do the heavy lifting of
looking of Dagger elements, these extensions are solely focused on how
the information is shown and used in user-facing scenarios.

This change is a bit lighter on tests than I'd like, because the core
extension implementations require the index to be enabled. Current tests
validate utility functions where appropriate; further tests will be
added when the index is turned on. (Additionally, all the existing tests
from v1 of Dagger support are generally more e2e and will cover all of
this logic.)

Bug: 265846405
Test: Included
Change-Id: I065291abd6013c7ad60b2e3fb9f1beaa742c5049

GitOrigin-RevId: 8f557a0b126d3fcdf4f727a6ce768d6042ea63a3