Enable dagger index
This change registers DaggerIndex so that it will start running and
building. The version of Dagger tooling that relies on the index isn't
turned on by default with this change, as it's still behind a flag; but
now it *can* be turned on with the flag and will work, since the index
will be built.

Bug: 265846405
Test: Included
Change-Id: I123428628a7ec175adefbc7fe3c70bb0a9669413

GitOrigin-RevId: fe596d0e478b4d329a6676286461d1f617b4c0a3