Add (Sub)Component and Module types
These types aren't used yet, but this clears the way for adding a
concept that does use them without the cruft of adding these various
strings.

This logic (especially the lists of strings and long when statements) is
being refactored in a subsequent change, where most of that logic gets
pushed into the "Concept" files so it can live directly with the logic
that understands it.

Bug: 265846405
Test: Included
Change-Id: Ib3b104fdd654a071a2bb6c293524ef8e34dba8db

GitOrigin-RevId: e5b747998840894778d0ebc9b011479250d19924