Rename ResourceRepositoryManage to StudioResourceRepositoryManager
Rendering pipeline currently depends on ResourceRepositoryManager. In an effort to make rendering pipeline independent of studio, we are extracting an interface from ResourceRepositoryManager as it
seems impossible to reuse it because the current implementation
heavily depends on studio-specific classes like
AndroidFacet.
As a first step, we are renaming ResourceRepositoryManager to StudioResourceRepositoryManager. In the next steps, we will extract out an interface and use that in the rendering pipeline.

Bug: 267600049
Test: StudioResourceRepositoryManagerTest
Change-Id: I4aa01c6013767c2060178c960bbb9124840c707d

GitOrigin-RevId: 1df4323f76f7d2fd1e1341152928dc0f5d75d625