Remove SamplesVariantRule
This rule is used for converting pre-Gradle metadata sample libraries to
a form where samples will be matched by Gradle.

This change updates the tests and published libraries used within to
contain Gradle metadata. No libraries are currently published with
samples without Gradle metadata and it seems they won't be in the future
either.

Bug: N/A
Test: tests adjusted
Change-Id: I0507dc006017e6dad9a3224199bef1d275ad0875

GitOrigin-RevId: d7d2dc32934446c1fa6f32724176d0a65620c13b