Remove parens for optional args completion
When Compose autocompletion fills in a function with (1) a required, final
@Composable argument and (2) only optional arguments otherwise, the
completion looks as follows:

```
Row(<caret>) {

}
```

According to b/181649371#comment4, this was part of the original design
for Compose autocompletion. But even the discussion in that bug shows
that this is behavior that doesn't adhere to users' expectation, and
other bugs like b/182564317 also expect that the parentheses should not
be present.

Note also that the default auto-complete behavior for non-Composable
functions would skip parentheses in a similar situation. The primary
thing that Compose autocomplete adds is parameter names (which are more
common and idiomatic in Compose than non-Compose), but the addition of
the parentheses for optional args doesn't add a lot to that goal.

This change brings Compose autocompletion in line with non-Compose
autocompletion as far as optional arguments are concerned, so that the
above completion would instead give:

```
Row {
  <caret>
}
```

Fixes: 181649371
Test: ComposeCompletionContributorTest
Change-Id: I03b289a424a4d6a24fd5bc1f2edce92cca08da6d

GitOrigin-RevId: e0501bbe73a234abd64fac3cc8a24ec9378a1e91