Improve Android typedef completion
Android TypeDefs (@IntDef/@StringDef) are not-quite-enum concepts that
complicate code completion. This change makes the IDE more aware of
these concepts and re-prioritizes and re-decorates completion results
for these types when they match the parameter being completed, in a
method call or class constructor.

If it takes too long to generate these completion results, or if they
are not generated at all, they may still not show up very highly. It
is left to future work to solve those issues.

Bug: 266854535
Test: TypeDefCompletionContributorTest.kt
Change-Id: Iefaa2c1604e231cb2d157e8c29a048bd40faf937

GitOrigin-RevId: c9079a43a167b0f2ebebda35da4d4e002be12da9