Fix assistant tutorial editor context menu
Only consume and go down the customized path if event comes from the
left mouse click. This is to prevent right clicks from removing the
selection from the editor, which is used by the context menu provider
to determine whether to show context menu items.

Also only provide the "Copy" action to the context menu.

Fixes: 247040492
Test: manual
Change-Id: I7ed5e6f9ad92126fbc63adb3684890d1659d0cd9

GitOrigin-RevId: c944499b192e5b184dc8750e548490c8efbc9b3e