Support refreshs on TutorialCards in Assistants
In order to have dynamic content in our bundles, we need a way to
refresh the bundles which will pull from the tutorials list of steps.
These lists of steps could change if the assistant offers different
steps depending on which ones are relevant to the user. Additionally,
add the ability to set the position to a particular step when doing
the refresh so we have the ability to provide a continuous experience.

Fixes: 263436530
Change-Id: Ic8fbe2d74c9c7575652bfc95cf291339a58b23ff

GitOrigin-RevId: 26b6e4663c45aa09f833cadea861636f4d2c4d67