AndroidProjectTaskRunner runs only android module related tasks
We want to avoid to run non-android modules tasks that end up
generating incorrect tasks names specially because on android
by default the Gradle tasks list isn't available to be able to
validate instead of "guessing" them. Something that IntelliJ
are using CachedModuleDataFinder.findAll(ProjectKeys.TASK)
to ensure the tasks are valid.

For this reason, those are going to be delegated to
GradleProjectTaskRunner, that knows how to handle ios or
pure java modules.

In addition, this change is aligned with IntelliJ changes
in this method for the same reason.

Bug: 270301645
Test: AndroidProjectTaskRunnerTest
Change-Id: Id174bf76fe62b4892420d9e5d1ad808982b2dece

GitOrigin-RevId: 997145c6cde0255ecc9a0da822ecab35f5ad00b0