[toml] Rename a banner type and class
Originally there was some potential for generalizing the banner from
"version catalogs" to "some feature of this project is too complicated
for the IDE to be confident in its understanding.  That possibility
has not materialized, and these banners are likely to be deleted
before any expectation of detecting other types of complexity.

(Leave the property name relating to hiding the banner as it was, so
that existing users are not affected.)

Bug: 267644534
Test: existing (no change in functionality)
Change-Id: Iebff48e7fa76c3e7f6542216766517f32fcfcdf0

GitOrigin-RevId: 099d1dea5a2186b6c6ef03c110888489a3c5859b