Guarding against NPE in AttributeProcessingUtil
The `getAdditionalAttributesClasses` method uses the result of a method
that can return null, but seems to assume that the returned value always
exists. This change rectifies that error.

Note that this appears to be somewhat old code, and the method maybe
used to return a collection with more than zero or one values. Since
that's no longer the case and the method is only used in one place, I'm
modifying it to return just a single value.

Fixes: 269325492
Test: n/a
Change-Id: I2176e8d698c4eec472923aaa566c559841886367

GitOrigin-RevId: 529bcc752bb19ca272daf1828c6252cd081906c7