Ensure correct order of clinit methods
In some scenarios, if AndroidSourceType subclasses were
accessed directly, their clinit would start running. JVM
then needs to run clinit for the superclass. In clinit
of AndroidSourceType, list of all subclasses is built.

However, because clinit of subclass is not done yet,
some of the static fields are null, and array ends up
containing null values.

This change ensures that list of all subclasses is
created outside of the clinit  of AndroidSourceType.

Bug: 270937899
Test: manual
Change-Id: I7d76ae6a7cc18cb75652caf9c5ea760a6eadbc62

GitOrigin-RevId: d2541dbaf3cb235ab2094f59651f95028d142fba