Improve disposables logic in downloads info UI
This UI subscribes to events from build and uses build finished
disposable to unsubscribe from it. In case UI itself gets disposed
it also needs to unsubscribe.
Before this change it was done via disposing buildFinishedDisposable
which is wrong because this instance can be shared with other
interested parties.
Instead, create another intermediate disposable that will be disposed
in case of build complete from parent buildCompleteDisposable and also
in case of UI disposal directly.

Test: existing + added test cases for disposing order
Bug: 265924450
Change-Id: Id37f7f6b318a5219bcc0747f3639f55a6016437f

GitOrigin-RevId: 711358b3cee3f2d0317d665b8a343e56b71ecd38