Move desugarLibraryConfigFiles to the module system
They are only reported in the models of Gradle
projects that have the feature enabled, and
therfore there isn't a single answer for
the intellij project.

The context that needs this information has a
module anyway, so it seems fine to just move
it to the Module system from the project system.

Bug: 262468597
Test: No semantic change for existing tests.
Change-Id: Icb4b6cffee45014e72d4ddd6db8e78891dab3997

GitOrigin-RevId: 47843e5ab606483fc2cd4399e24b78b9b86c7af3