Detach Blaze execution infrastructure from old one
Make a copy of LaunchTaskRunner for Blaze - BlazeAndroidConfigurationExecutor,
this executor is dedicated BlazeRunConfiguration.

Create copy of all interfaces LaunchTasksProvider, LaunchTask and so on
for Blaze - BlazeLaunchTask, BlazeLaunchTasksProvider etc.
For now this new interfaces live in core module, but should the file
should be moved to blaze module as soon as it supports Kotlin.

Bug: 269755515
Test: manually run/debug app, run debug instrumented tests
Change-Id: I6622d9259706df30b3e98471bf947fbd470ca105

GitOrigin-RevId: f32354f98b0164ef75220e0ab2702184a8c58aee