Refactor related items to allow customization
When resolving related Dagger items, the current infrastructure assumes
that we will always look up items in the index, filter to some set of
known types, and be done. In reality, some related items may need to be
looked up in a different way (eg, Modules included in a Component aren't
in the index, and instead are found directly on the Component
annotation.)

To accomplish this, I've made DaggerElement abstract and let derived
implementations specify how they find related elements. This change is
just refactoring and doesn't change behavior; a subsequent change will
add different kinds of related item resolution.

Test: Included
Bug: 265846405
Change-Id: Ie04bfa3f7893846d00d98373d18a114e9cbdac85

GitOrigin-RevId: 35e7ac0b5debaf6a784b86ebbe4c1ac0205010fd