K2 ComposeCreateComposableFunction Quick Fix
Introduce K2 version of CreateComposableFunction quick fix.

The changes splits the tree of utility functions supporting
the functionality in order to avoid reanalysis of Kotlin code.

TODO(b/267429486): The new functionality is limited in scope
as `CreateCallableFromUsage` is not yet supported on K2.
Revisit `ComposeCreateComposableFunctionQuickFix` after adding
K2 version of `CreateCallableFromUsage`.

Bug: 267429486, 269680473
Test: bazel test //tools/adt/idea/compose-ide-plugin:intellij.android.compose-ide-plugin.tests_tests \
      --test_filter=ComposeUnresolvedFunctionFixContributorTest \
      --jvmopt='-Didea.kotlin.plugin.use.k2=true'
Change-Id: Iaddd5e08cbdd722d49332e7a276b0d78b9003192

GitOrigin-RevId: 9108b0c3d5a888de6943a59757f276ad61438dce